+++ pkg/util/http_test.go
@@ -46,7 +46,7 @@ func TestHexStringToTraceID(t *testing.T) {
 
 	for _, tt := range tc {
 		t.Run(tt.id, func(t *testing.T) {
-			actual, err := hexStringToTraceID(tt.id)
+			actual, err := HexStringToTraceID(tt.id)
 
 			if tt.expectError != nil {
 				assert.Equal(t, tt.expectError, err)